odmx: replace ssi.general with odmx.file_utils (which overlaps with ssi.general but has only the functions used in the odmx pipeline
